terminado paginación y corregido los errores de relaciones

+++ backend/main/resources/BolsonPendiente.py
@@ -4,34 +4,56 @@ from .. import db
 from main.models import BolsonModel
 from datetime import datetime
 
-BOLSONESPENDIENTES = {
-    1: {'name': 'bolsonpendiente1'},
-    2: {'name': 'bolsonpendiente2'},
-    3: {'name': 'bolsonpendiente3'}
-}
 
 class BolsonPendiente(Resource):
     def get(self,id):
         bolsonpendiente = db.session.query(BolsonModel).get_or_404(id)
-        return bolsonpendiente.to_json()
+        if bolsonpendiente.aprobado == 0:
+            return bolsonpendiente.to_json()
+        else:
+            return '', 404
     def delete(self,id):
         bolsonpendiente = db.session.query(BolsonModel).get_or_404(id)
-        db.session.delete(bolsonpendiente)
-        db.session.commit()
-        return '', 204
+        if bolsonpendiente.aprobado == 0:
+            db.session.delete(bolsonpendiente)
+            db.session.commit()
+            return '', 204
+        else:
+            return '', 404
     def put(self,id):
         bolsonpendiente = db.session.query(BolsonModel).get_or_404(id)
         data = request.get_json().items()
         for key, value in data:
             setattr(bolsonpendiente, key, value)
-        db.session.add(bolsonpendiente)
-        db.session.commit()
-        return bolsonpendiente.to_json(), 201
+        if bolsonpendiente.aprobado == 0:
+            db.session.add(bolsonpendiente)
+            db.session.commit()
+            return bolsonpendiente.to_json(), 201
+        else:
+            return '', 404
 
 class BolsonesPendientes(Resource):
     def get(self):
-        bolsones = db.session.query(BolsonModel).all()
-        return jsonify([bolson.to_json() for bolson in bolsones])
+        page = 1
+        per_page = 10
+        bolsones = db.session.query(BolsonModel).filter(BolsonModel.aprobado == 0)
+
+        if request.get_json():
+            filters = request.get_json().items()
+            for key, value in filters:
+                if key =="page":
+                    page = int(value)
+                if key =="per_page":
+                    per_page = int(value)
+        bolsones = bolsones.paginate(page, per_page, True, 30)
+
+        return jsonify({
+            'bolsonespendientes': [bolson.to_json() for bolson in bolsones.items],
+            'total': bolsones.total,
+            'pages': bolsones.pages,
+            'page': page
+        })
+
     def post(self):
         bolson = BolsonModel.from_json(request.get_json())
         db.session.add(bolson)
